CE-2367 replace method isTitleWithNonportableInfobox
