+++ extensions/wikia/TemplateDraft/TemplateDraftHooks.class.php
@@ -1,4 +1,5 @@
 <?php
+use \Wikia\PortableInfobox\Helpers\PortableInfoboxClassification;
 
 class TemplateDraftHooks {
 
@@ -75,8 +76,9 @@ class TemplateDraftHooks {
 	 * @return bool
 	 */
 	public static function onEditPageLayoutShowIntro( &$preloads, Title $title ) {
+		global $wgEnablePortableInfoboxExt;
 		if ( $title->getNamespace() == NS_TEMPLATE ) {
-			$model = InsightsHelper::getInsightModel( InsightsUnconvertedInfoboxesModel::INSIGHT_TYPE );
+			$contentText = ( new WikiPage( $title ) )->getText();
 			if ( TemplateDraftHelper::isTitleDraft( $title ) ) {
 				$base = Title::newFromText( $title->getBaseText(), NS_TEMPLATE );
 				$baseHelp = Title::newFromText( 'Help:PortableInfoboxes' );
@@ -95,23 +97,24 @@ class TemplateDraftHooks {
 							wfMessage( 'templatedraft-module-view-parent' )->plain() )
 					)->escaped(),
 				];
-			} elseif ( !TemplateDraftHelper::titleHasPortableInfobox( $title )
-				&& isset( $model->fetchArticlesData()[$title->getArticleID()] ) ) {
-				$draft = wfMessage( 'templatedraft-subpage' )->inContentLanguage()->escaped();
-				$base = Title::newFromText( $title->getBaseText() .'/'. $draft, NS_TEMPLATE );
-				$draftUrl = $base->getFullUrl( [
-					'action' => 'edit',
-					TemplateConverter::CONVERSION_MARKER => 1,
-				] );
-				$preloads['EditPageIntro'] = [
-					'content' => wfMessage( 'templatedraft-module-editintro-please-convert' )->rawParams(
-						Xml::element( 'a', [
-							'href' => $draftUrl,
-							'target' => '_blank'
-						],
-							wfMessage( 'templatedraft-module-button-create' )->plain() )
-					)->escaped(),
-				];
+			} elseif ( !TemplateDraftHelper::titleHasPortableInfobox( $title ) && $wgEnablePortableInfoboxExt ) {
+				if ( PortableInfoboxClassification::isTitleWithNonportableInfobox( $title->getText(), $contentText ) ) {
+					$draft = wfMessage( 'templatedraft-subpage' )->inContentLanguage()->escaped();
+					$base = Title::newFromText( $title->getBaseText() . '/' . $draft, NS_TEMPLATE );
+					$draftUrl = $base->getFullUrl( [
+						'action' => 'edit',
+						TemplateConverter::CONVERSION_MARKER => 1,
+					] );
+					$preloads['EditPageIntro'] = [
+						'content' => wfMessage( 'templatedraft-module-editintro-please-convert' )->rawParams(
+							Xml::element( 'a', [
+								'href' => $draftUrl,
+								'target' => '_blank'
+							],
+								wfMessage( 'templatedraft-module-button-create' )->plain() )
+						)->escaped(),
+					];
+				}
 			}
 		}
 		return true;
